+++ apps/community/src/apis/lab/queries.ts
@@ -1,0 +1,14 @@
+import LabService from './lab-service';
+
+const queryKeys = {
+  all: ['labs'] as const,
+};
+
+const labsQueryOptions = {
+  all: () => ({
+    queryKey: queryKeys.all,
+    queryFn: () => LabService.getLabs(),
+  }),
+};
+
+export { labsQueryOptions };
