+++ apps/community/src/types/lab.ts
@@ -1,0 +1,10 @@
+interface Lab {
+  id: number;
+  name: string;
+  location: string;
+  site: string;
+  professor: string;
+  img?: string;
+}
+
+export type { Lab };
